# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storage Application API",
        "description": "Storage API",
        "contact": {
            "name": "Min",
            "url": "https://www.facebook.com/Min9802/",
            "email": "hotro.minservice@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/storage/client/file/get": {
            "post": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "get file",
                "description": "use token",
                "operationId": "get file",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "path": {
                                        "description": "file path to save",
                                        "type": "string",
                                        "example": "Admin/image/redis.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "id": 1,
                                    "name": "jx3m8NPBgy.png",
                                    "type": "image",
                                    "path": "public/image/jx3m8NPBgy.png",
                                    "status": 0,
                                    "created_at": "2022-09-29T17:03:15.000000Z",
                                    "updated_at": "2022-09-29T17:13:16.000000Z",
                                    "deleted_at": null,
                                    "url": "http://domain.com/storage/image/jx3m8NPBgy.png",
                                    "pivot": {
                                        "user_id": 1,
                                        "file_id": 1
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "notfound"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/storage/client/folder/list": {
            "get": {
                "tags": [
                    "Storage Api folder"
                ],
                "summary": "get folder list",
                "description": "use token",
                "operationId": "folderlist",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "message": "Get data success",
                                    "folders": [
                                        "image",
                                        "audio"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "getdata fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/folder/exist": {
            "post": {
                "tags": [
                    "Storage Api folder"
                ],
                "summary": "Check folder exist",
                "description": "use token",
                "operationId": "folderexists",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name folder",
                                        "type": "string",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "message": "Folder Exists"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/folder/create": {
            "post": {
                "tags": [
                    "Storage Api folder"
                ],
                "summary": "Add new folder",
                "description": "use token",
                "operationId": "createfolder",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name folder",
                                        "type": "string",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "message": "Add success",
                                                "folder": "image"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": {
                                                "message": "Folder exists"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Add folder fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/folder/rename": {
            "post": {
                "tags": [
                    "Storage Api folder"
                ],
                "summary": "Rename a folder",
                "description": "use token",
                "operationId": "renamefolder",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name folder",
                                        "type": "string",
                                        "example": "image"
                                    },
                                    "newname": {
                                        "description": "name folder",
                                        "type": "string",
                                        "example": "images"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "message": "Rename success",
                                                "folder": "image"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": {
                                                "message": "Folder not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rename folder fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/folder/getfile": {
            "post": {
                "tags": [
                    "Storage Api folder"
                ],
                "summary": "Get file in folder",
                "description": "use token",
                "operationId": "getfilefolder",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name folder",
                                        "type": "string",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "message": "get data success",
                                                "files": [
                                                    {
                                                        "id": 3,
                                                        "name": "redis.png",
                                                        "folder": "image",
                                                        "type": "image/png",
                                                        "hash": "redis.png",
                                                        "path": "Admin/image/redis.png",
                                                        "url": "http://127.0.0.1:8001/storage/Admin/image/redis.png",
                                                        "status": 0,
                                                        "created_at": "2023-03-02T17:30:13.000000Z",
                                                        "updated_at": "2023-03-02T17:30:13.000000Z",
                                                        "deleted_at": null,
                                                        "pivot": {
                                                            "user_id": "1",
                                                            "file_id": 3
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": {
                                                "message": "Folder not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "get data fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "FileSystem": {
                "title": "FileSystem",
                "description": "FileSystem model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "file_name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "pivot": {
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "FileSystem"
                }
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8001/api/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8001/api/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8001/api/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}